// packages/boilerplate/src/wheelroom/elements/types/media.ts
export interface MediaObject {
  title: string
  description: string
  fluid?: {
    base64: any
    aspectRatio: any
    src: string
    srcSet: any
    sizes: any
  }
  file?: {
    url: string
    fileName: string
    contentType: string
    details: {
      image: {
        height: number
        width: number
      }
    }
  }
}